refactored listenTo
